refactor: merge seo sitemap and render folders
